Move generation related enumerations into common header to simplify includes
