+++ src/hotspot/share/gc/shenandoah/heuristics/shenandoahHeuristics.hpp
@@ -25,7 +25,6 @@
 #ifndef SHARE_GC_SHENANDOAH_HEURISTICS_SHENANDOAHHEURISTICS_HPP
 #define SHARE_GC_SHENANDOAH_HEURISTICS_SHENANDOAHHEURISTICS_HPP
 
-#include "gc/shenandoah/shenandoahHeap.hpp"
 #include "gc/shenandoah/shenandoahPhaseTimings.hpp"
 #include "gc/shenandoah/shenandoahSharedVariables.hpp"
 #include "memory/allocation.hpp"
@@ -57,6 +56,7 @@
 
 class ShenandoahCollectionSet;
 class ShenandoahHeapRegion;
+class ShenandoahGeneration;
 
 class ShenandoahHeuristics : public CHeapObj<mtGC> {
   static const intx Concurrent_Adjust   = -1; // recover from penalties
